requirements.txt
altair
pandas
streamlit
db-dtypes
google-cloud-bigquery
plotly.express